requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
apipkg==1.5
attrs==19.3.0
execnet==1.7.1
importlib-metadata==1.6.0
more-itertools==8.2.0
mysql-connector-python==8.0.20
packaging==20.3
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-forked==1.1.3
pytest-xdist==1.32.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
